Rename source-repos to site-data.
